Add initial unit tests of block statement in attribute replacement
