+++ tests/unit.ts
@@ -1,4 +1,5 @@
 import { camelizePropName } from '../src/styles'
+import { preProcessUnsupportedParserFeatures } from '../src/unsupportedParserFeatures';
 
 const mockStyles
   // eslint-disable-next-line
@@ -11,3 +12,21 @@ describe('styles', () => {
     expect(camelizePropName('-webit-transition')).toBe('WebitTransition')
   })
 })
+
+describe('preProcessUnsupportedParserFeatures', () => {
+  describe('when no block statement exists in attributes', () => {
+    [
+      '<div><div>',
+      '<div>{{name}}<div>',
+      '<div>{{#if isTrue}}text{{/if}}<div>',
+      '<div>{{#if isTrue}}{{name}}{{/if}}<div>',
+      '<div id={{id}}><div>',
+      '<div id="{{id}}"><div>'
+    ].forEach((template) => {
+      test('should return template unchanged', () => {
+        const result = preProcessUnsupportedParserFeatures(template);
+        expect(result).toEqual(template);
+      });
+    });
+  });
+});
